requirements.txt
#base packages
pandas
python-dotenv
openpyxl

#For building RAG
langchain
langchain-community
langchainhub
langchain-chroma
langchain_openai
langchain_core
langchain_text_splitters

beautifulsoup4

#For connecting with AstraDB
astrapy

#For vector embedding
openai

#Inquirer
inquirer

#Metrics for evaluating RAGs
asyncio #For asyncronuos programming
sacrebleu
rouge_score
ragas